==> src/03_preprocessing_0.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("D:\Kuliah\Semester 7\TA\Fraud_Insured_Automobile_Detection\data\df_feature_engineered.csv")

# Preprocessing tahap Awal

for col in df.columns:
    print(f"Column '{col}':")
    print(df[col].unique())
    print("\n" + "-"*50 + "\n")

print(df['umbrella_limit'].unique())

# mengubah nilai minus di umbrella_limit jadi plus
df['umbrella_limit'] = df['umbrella_limit'].abs()

# convert umbrella_limit ke categorical
df['umbrella_limit'] = df['umbrella_limit'].astype('object')

# convert policy_deductable ke categorical
df['policy_deductable'] = df['policy_deductable'].astype('object')

#imputasi missing value kolom authorities_contacted dengan 'UNKNOWN'
df['authorities_contacted'] = df['authorities_contacted'].fillna('UNKNOWN')

# Ganti '?' dengan NaN
cols_replace = ['property_damage', 'police_report_available', 'collision_type']
for c in cols_replace:
    if c in df.columns:
        df[c] = df[c].replace('?', np.nan)

# Drop kolom yang tidak relevan / terlalu unik
df = df.drop(['policy_bind_date', 'incident_date', 'policy_number', 'insured_zip',
              'incident_location', 'auto_model', 'auto_make'], axis=1)

# Encode target
df['fraud_reported'] = df['fraud_reported'].map({'N': 0, 'Y': 1})

# save output
df.to_csv("D:\Kuliah\Semester 7\TA\Fraud_Insured_Automobile_Detection\data\df_clean.csv", index=False)
