==> src/04_split_data.ipynb <==
from sklearn.model_selection import train_test_split
import pandas as pd
import numpy as np

df = pd.read_csv("D:\Kuliah\Semester 7\TA\Fraud_Insured_Automobile_Detection\data\df_clean.csv")

# Split Data

# PEMISAHAN FITUR DAN TARGET
X = df.drop('fraud_reported', axis=1)
y = df['fraud_reported']

# Menambahkan seed untuk reproducibility
seed = 42
# Split data dengan stratifikasi
X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.3, random_state=seed, stratify=y
)

import pickle

with open("train_test_data.pkl", "wb") as f:
    pickle.dump(
        {
            "X_train": X_train,
            "X_test": X_test,
            "y_train": y_train,
            "y_test": y_test
        },
        f
    )
